Fix function_clause when latest rev and the flag are explicitly provided
Current implementation assumes that amount of received responses will
match amount of specified revisions for open_revs query parameter.
However, for case when open_revs contains old and leaf revisions with
among with latest=true query parameter that's not true: RawReplies will
only contain the latest revision, while old ones will not be included.
This makes All and RawReplies lists mismatch by the length making
zipwith/3 crash with the function_clause error.

The fix respects both optimistic and pessimistic cases: when all is fine
we use zipwith/3 as we did before, but otherwise we're trying to group
them only by leaf revisions that document have and which are defined in
request parameter.

COUCHDB-2863